baidu ue

+++ mserver/src/main/java/com/co/example/dao/system/TMenuDao.java
@@ -3,5 +3,5 @@ package com.co.example.dao.system;
 import com.co.example.common.dao.BaseDao;
 import com.co.example.entity.system.TMenu;
 
-public interface TMenuDao extends BaseDao<TMenu, Integer> {
+public interface TMenuDao extends BaseDao<TMenu, Long> {
 }
