+++ mserver/src/main/java/com/co/example/dao/system/impl/TMenuDaoImpl.java
@@ -6,5 +6,5 @@ import com.co.example.entity.system.TMenu;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public class TMenuDaoImpl extends BaseDaoImpl<TMenu, Integer> implements TMenuDao {
+public class TMenuDaoImpl extends BaseDaoImpl<TMenu, Long> implements TMenuDao {
 }
